sema: move symbol related stuff to separate module
Currently everything is just in one file, but it's probably best to split it up more
